Extract deleteBlankLines to external function
